% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{constant_cov}
\alias{constant_cov}
\title{constant_cov}
\usage{
constant_cov(x, y, sigma)
}
\arguments{
\item{x}{numerical vector}

\item{y}{numerical vector}

\item{sigma}{numerical value}
}
\value{
constant covariance of input parameters
}
\description{
constant_cov
}
\examples{
x <- runif(3,-1,1)
y <- runif(3,-1,1)
constant_cov(x,y,1)
}
