% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_2.1.R
\name{cov_cross}
\alias{cov_cross}
\title{calculating the variance-covariance-matrix}
\usage{
cov_cross(x, y, cov, ...)
}
\arguments{
\item{x}{first list of input vectors}

\item{y}{second list of input vectors}

\item{cov}{covariance function, that returns a scalar}

\item{...}{further parameters of cov, they have to be named.}
}
\value{
variance-covariance-matrix of dimension 'length(x)':'length(y)'
}
\description{
calculating the variance-covariance-matrix
}
\examples{
cov_cross(list(1,2,3), list(1,2), cov = function(x,y) x*y)

}
