% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{covariance_list}
\alias{covariance_list}
\title{covariance_list}
\usage{
covariance_list(
  l = NULL,
  alpha = NULL,
  gamma = NULL,
  x = NULL,
  y = NULL,
  sigma = NULL,
  tibble = FALSE
)
}
\arguments{
\item{l}{positive double value}

\item{alpha}{positive double value}

\item{gamma}{double value between 0 and 2}

\item{x}{numerical vector}

\item{y}{numerical vector}

\item{sigma}{numerical vector or value}

\item{tibble}{logical value}
}
\value{
list of different covariance of given input parameters, optionally as tibble
}
\description{
covariance_list
}
