% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_laplace.R
\name{diag_block_matrix}
\alias{diag_block_matrix}
\title{Building block matrices}
\usage{
diag_block_matrix(matrix_list)
}
\arguments{
\item{matrix_list}{list of quadratic matrices}
}
\value{
one diagonal block matrix
}
\description{
Building block matrices
}
\examples{
l <- list(
  matrix(1:9, nrow = 3),
  diag(4),
  diag(c(1,2,3,3,3)),
  matrix(90:98, nrow = 3)
)

diag_block_matrix(l)


}
