% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{exp_cov}
\alias{exp_cov}
\title{exp_cov}
\usage{
exp_cov(x, y, l)
}
\arguments{
\item{x}{numerical vector}

\item{y}{numerical vector}

\item{l}{positive double value}
}
\value{
exponential covariance of input parameters
}
\description{
exp_cov
}
\examples{
exp_cov(c(1,1),c(2,1), 4)
}
