% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_laplace.R
\name{find_mode_mc_laplace}
\alias{find_mode_mc_laplace}
\title{Mode-finding for multiclass laplace approximation}
\usage{
find_mode_mc_laplace(K_list, y)
}
\arguments{
\item{K_list}{list of of variance-covariance matrices. Each class label gets
it's own matrix}

\item{y}{vector of resulting class labels, it's values are either 0 or 1.}
}
\value{
list of two elements:
- mode: numerical vector with mode function values
- log_marginal_likelihood: value of the logarithmic marginal likelihood
}
\description{
Mode-finding for multiclass laplace approximation
}
\details{
Assuming n points, for which the right labeling is known, and C possible
classes, than this vector is build up like this:

- for each class, we consider the vector y_c, which has length n and whose
entries are defined as 1, if point i belongs to class c, or 0 otherwise.

- y is created by joining all y_c - vectors.
}
\examples{
#Assuming 2 classes and 3 learning points
K_list <- list(
  diag(c(1,2,3)),
  diag(c(0.4,2.3,1))
)

y <- c(0,0,1,1,1,0)

find_mode_mc_laplace(K_list, y)
}
