% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{gamma_exp_cov}
\alias{gamma_exp_cov}
\title{gamma_exp_cov}
\usage{
gamma_exp_cov(x, y, l, gamma)
}
\arguments{
\item{x}{numerical value}

\item{y}{numerical value}

\item{l}{positive double value}

\item{gamma}{double value between 0 and 2}
}
\value{
gamma exponential covariance of input parameters
}
\description{
gamma_exp_cov
}
\examples{
gamma_exp_cov(c(1,1),c(2,1), 4, 2/3)
}
