% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_cov}
\alias{get_cov}
\title{Getting the used covariance function}
\usage{
get_cov(obj)
}
\arguments{
\item{obj}{a instance of class "gp"}
}
\value{
a list of two elements:
name: the name of the covariance function
cov_fun: the closure used to calculate the values.
it's arguments are numerical vectors x and y, both of the same length like
the input_dimension of the used gp-object.
}
\description{
Getting the used covariance function
}
\examples{
p <- new.gp() \%>\%
     add_data(1:3, 1:3)
f <- get_cov(p)
f$cov_fun(12, 34)

}
\seealso{
set_cov for setting the covariance function of the gp-object
}
