% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_cov_name}
\alias{get_cov_name}
\title{Getting the name of the used covariance function}
\usage{
get_cov_name(obj)
}
\arguments{
\item{obj}{a instance of class "gp"}
}
\value{
character vector, the name of the used covariance function
}
\description{
Getting the name of the used covariance function
}
\details{
For further information about usable covariance functions, you
might consider the documentation of "set_cov"
}
\examples{
library(dplyr) #to include pipe-operator
p <- new.gp() \%>\%
     add_data(1:10,1:10)
get_cov_name(p)
#outputs default value: "squared_exp"

#------------------------------------------
p <- new.gp(cov_fun = "linear")
get_cov_name(p)
#output: "linear"
}
\seealso{
set_cov() for setting the covariance function
}
