% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_data}
\alias{get_data}
\title{Getting the data of a gp}
\usage{
get_data(obj, df = T)
}
\arguments{
\item{obj}{a instance of class "gp"}

\item{df}{a logical value, whether the output should be a data.frame (df = T)
or as a listed object.}
}
\value{
either a data.frame or a list of points
}
\description{
Getting the data of a gp
}
\examples{
library(dplyr) #to include pipe-operator

p1 <- new.gp() \%>\% add_data(matrix(c(1:10, 2:11, 3:12),nrow=3), 1:10)
get_data(p1)
#output is a data.frame

get_data(p1, FALSE)
#output is a list of listed points

}
