% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_K}
\alias{get_K}
\title{Getting the variance-covariance matrix}
\usage{
get_K(obj)
}
\arguments{
\item{obj}{a instance of class "gp"}
}
\value{
variance-covariance matrix of "obj"
}
\description{
Getting the variance-covariance matrix
}
\examples{
library(dplyr) #to include pipe-operator
#initializing a new 'gp' object
p <- new.gp() \%>\%
     add_data(1:3, 1:3)

get_K(p)


}
