% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_log_marginal_likelihood}
\alias{get_log_marginal_likelihood}
\title{Getting the value of the logarithmic marginal likelihood}
\usage{
get_log_marginal_likelihood(obj)
}
\arguments{
\item{obj}{a instance of class "gp"}
}
\value{
numerical value of the logarithmic marginal likelihood
}
\description{
Getting the value of the logarithmic marginal likelihood
}
\details{
The marginal likelihood of the posteriori according to the
data added to the gp object. This value is highly depending on the
covariance function and the parameters, that are used. By maximizing
this value, you get one possible approach of optimzing the hyperparameters
of the used Gaussian Process model. If you are interested in further
informations about optimizing hyperparameters, see also optimize_gp
}
\examples{
p <- new.gp() \%>\%  add_data(1:10, 1:10)

get_log_marginal_likelihood(p)

}
