% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_mean_fun}
\alias{get_mean_fun}
\title{Getting the mean function of a gp object}
\usage{
get_mean_fun(obj)
}
\arguments{
\item{obj}{a instance of class "gp"}
}
\value{
a closure, that returns the prior mean value
}
\description{
Getting the mean function of a gp object
}
\examples{
p1 <- new.gp()
f <- get_mean_fun(p1)

f(c(1,2,3))
# equals zero per default

}
\seealso{
set_mean_fun for further information about the mean function
}
