% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_noise}
\alias{get_noise}
\title{Getting the value of the noise}
\usage{
get_noise(obj)
}
\arguments{
\item{obj}{a instance of class "gp"}
}
\value{
value of the variance of the noise according to the
data of the "gp" object
}
\description{
Getting the value of the noise
}
\examples{
p <- new.gp()

get_noise(p)
#default value

}
\seealso{
set_noise() for setting this value
}
