% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_parameter}
\alias{get_parameter}
\title{Getting parameters of a gp instance}
\usage{
get_parameter(obj, used = F)
}
\arguments{
\item{obj}{a instance of class "gp"}

\item{used}{logical vector, whether just the parameters used by the current
covariance function should be returned, FALSE per default}
}
\value{
named list of numerical vectors
}
\description{
Getting parameters of a gp instance
}
\details{
If you are interested in the parameters that are used by the different
available covariance functions, visit the vignette of this package
}
\examples{
p1 <- new.gp() \%>\% add_data(matrix(c(1:10, 2:11, 3:12),nrow=3), 1:10)

get_parameter(p1)
#outputs all parameters stored in the gp object

get_parameter(p1, TRUE)
#just outputs value of the parameter l
}
