% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_prediction}
\alias{get_prediction}
\title{Getting prediction using Gaussian Process Model}
\usage{
get_prediction(obj, x_input)
}
\arguments{
\item{obj}{a instance of class "gp"}

\item{x_input}{numerical vector, it's length has to equal the dimension of the
learning data of the gp object. It's not possible to handle more than one input
per function call.}
}
\value{
a list with three items:
f_pred: predicted value at position x_input
var_f : the variance of the result
log_marginal_likelihood: the log_marginal_likelihood of the gp object according
to the overall setup
}
\description{
Getting prediction using Gaussian Process Model
}
\examples{
library(dplyr) #to include pipe-operator

p <- new.gp() \%>\%  add_data(1:10, 1:10)
get_prediction(p,11)

}
\seealso{
predict_gauss(): if you do not want to use the class 'gp'
Notice: We really recommend using the 'gp' class due to efficiency in the case
of predicting more than one value.
}
