% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{gp}
\alias{gp}
\title{#Use add_data() instead}
\description{
#Use add_data() instead

#Use add_data() instead
}
\examples{

## ------------------------------------------------
## Method `gp$plot`
## ------------------------------------------------

p <- new.gp()
p <- gaussprocess::add_data(p, 1:10,sin(1:10))
# not a sampled graph
plot(p, x_start = -3, x_end = 8)
# sampled graph
plot(p, x_start = -3, x_end = 8, sampled_graph = T)

#Printing a `gp` instance
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gp-add_data}{\code{gp$add_data()}}
\item \href{#method-gp-get_cov_name}{\code{gp$get_cov_name()}}
\item \href{#method-gp-get_K}{\code{gp$get_K()}}
\item \href{#method-gp-get_cov}{\code{gp$get_cov()}}
\item \href{#method-gp-get_data}{\code{gp$get_data()}}
\item \href{#method-gp-get_parameter}{\code{gp$get_parameter()}}
\item \href{#method-gp-get_mean_fun}{\code{gp$get_mean_fun()}}
\item \href{#method-gp-get_noise}{\code{gp$get_noise()}}
\item \href{#method-gp-get_prediction}{\code{gp$get_prediction()}}
\item \href{#method-gp-get_log_marginal_likelihood}{\code{gp$get_log_marginal_likelihood()}}
\item \href{#method-gp-get_input_dim}{\code{gp$get_input_dim()}}
\item \href{#method-gp-set_cov}{\code{gp$set_cov()}}
\item \href{#method-gp-set_parameter}{\code{gp$set_parameter()}}
\item \href{#method-gp-set_noise}{\code{gp$set_noise()}}
\item \href{#method-gp-set_mean_fun}{\code{gp$set_mean_fun()}}
\item \href{#method-gp-optim_parameter}{\code{gp$optim_parameter()}}
\item \href{#method-gp-plot}{\code{gp$plot()}}
\item \href{#method-gp-print}{\code{gp$print()}}
\item \href{#method-gp-update_marginal_likelihood}{\code{gp$update_marginal_likelihood()}}
\item \href{#method-gp-clone}{\code{gp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-add_data"></a>}}
\if{latex}{\out{\hypertarget{method-gp-add_data}{}}}
\subsection{Method \code{add_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$add_data(x, y, noise = 0.1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector}

\item{\code{y}}{numeric vector}

\item{\code{noise}}{numeric vector}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Use ?add_data() to get to the real-used function
#Use get_cov_name() instead
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-get_cov_name"></a>}}
\if{latex}{\out{\hypertarget{method-gp-get_cov_name}{}}}
\subsection{Method \code{get_cov_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$get_cov_name()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Use ?get_cov_name() to get to the real-used function
#Use get_K() instead
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-get_K"></a>}}
\if{latex}{\out{\hypertarget{method-gp-get_K}{}}}
\subsection{Method \code{get_K()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$get_K()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Use ?get_K() to get to the real-used function
#Use get_cov()
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-get_cov"></a>}}
\if{latex}{\out{\hypertarget{method-gp-get_cov}{}}}
\subsection{Method \code{get_cov()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$get_cov()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Use ?get_cov() to get to the real-used function
#Use get_data() instead
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-gp-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$get_data()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Use ?get_data() to get to the real-used function
#Use get_parameter() instead
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-get_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-gp-get_parameter}{}}}
\subsection{Method \code{get_parameter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$get_parameter()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Use ?get_parameter() to get to the real-used function
#Use get_mean_fun() instead
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-get_mean_fun"></a>}}
\if{latex}{\out{\hypertarget{method-gp-get_mean_fun}{}}}
\subsection{Method \code{get_mean_fun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$get_mean_fun()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Use ?get_mean_fun() to get to the real-used function
#Use get_noise() instead
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-get_noise"></a>}}
\if{latex}{\out{\hypertarget{method-gp-get_noise}{}}}
\subsection{Method \code{get_noise()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$get_noise()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Use ?get_noise() to get to the real-used function
#Use get_prediction() instead
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-get_prediction"></a>}}
\if{latex}{\out{\hypertarget{method-gp-get_prediction}{}}}
\subsection{Method \code{get_prediction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$get_prediction(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{numeric vector}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Use ?get_prediction() to get to the real-used function
#Use get_log_marginal_likelihood() instead
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-get_log_marginal_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-gp-get_log_marginal_likelihood}{}}}
\subsection{Method \code{get_log_marginal_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$get_log_marginal_likelihood()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Use ?get_log_marginal_likelihood to get to the real-used function
#getting input dimension
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-get_input_dim"></a>}}
\if{latex}{\out{\hypertarget{method-gp-get_input_dim}{}}}
\subsection{Method \code{get_input_dim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$get_input_dim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
input dimension
#Use set_cov() instead
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-set_cov"></a>}}
\if{latex}{\out{\hypertarget{method-gp-set_cov}{}}}
\subsection{Method \code{set_cov()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$set_cov(cov)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cov}}{name of the covariance function}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Use ?set_cov() to get to the real-used function
#Use set_parameter() instead
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-set_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-gp-set_parameter}{}}}
\subsection{Method \code{set_parameter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$set_parameter(
  sigma = NULL,
  l = NULL,
  alpha = NULL,
  sigma0 = NULL,
  gamma = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sigma}}{numeric vector}

\item{\code{l}}{numeric vector}

\item{\code{alpha}}{numeric vector}

\item{\code{sigma0}}{numeric vector}

\item{\code{gamma}}{numeric vector}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Use ?set_parameter to get to the real-used function
#Use set_noise() instead
}

\subsection{Returns}{
updated element
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-set_noise"></a>}}
\if{latex}{\out{\hypertarget{method-gp-set_noise}{}}}
\subsection{Method \code{set_noise()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$set_noise(noise)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{noise}}{numeric vector}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Use ?set_noise to get to the real-used function
#Use ?set_mean_fun instead
}

\subsection{Returns}{
updated element
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-set_mean_fun"></a>}}
\if{latex}{\out{\hypertarget{method-gp-set_mean_fun}{}}}
\subsection{Method \code{set_mean_fun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$set_mean_fun(f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{numeric vector or closure}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Use ?set_mean_fun to get to the real-used function
#Use optim_gp() instead
}

\subsection{Returns}{
updated element
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-optim_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-gp-optim_parameter}{}}}
\subsection{Method \code{optim_parameter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$optim_parameter()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
updated element
#Plotting a `gp`-instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-plot"></a>}}
\if{latex}{\out{\hypertarget{method-gp-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$plot(
  x_start = 0,
  x_end = 10,
  n_samples = 3,
  n_points = 100,
  sampled_graph = FALSE,
  plotly_obj = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_start}}{start value in x-direction}

\item{\code{x_end}}{end value in x-direction}

\item{\code{n_samples}}{if you want to plot a sampled graph, you can specify the number
of samplings, that should be done.}

\item{\code{n_points}}{number of points that are used to generate the plot}

\item{\code{sampled_graph}}{if a sampled graph should be plotted (TRUE) or not (FALSE)}

\item{\code{plotly_obj}}{if the return value should be a plotly object. FALSE per default}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
either a ggplot or a plotly object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{p <- new.gp()
p <- gaussprocess::add_data(p, 1:10,sin(1:10))
# not a sampled graph
plot(p, x_start = -3, x_end = 8)
# sampled graph
plot(p, x_start = -3, x_end = 8, sampled_graph = T)

#Printing a `gp` instance
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-print"></a>}}
\if{latex}{\out{\hypertarget{method-gp-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
print element
#Updating marginal likelihood
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-update_marginal_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-gp-update_marginal_likelihood}{}}}
\subsection{Method \code{update_marginal_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$update_marginal_likelihood()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
modified element
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
