% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_using_gp.R
\name{gp_app}
\alias{gp_app}
\title{Starting Shiny App to Work with GP-Objects}
\usage{
gp_app(obj = NULL)
}
\arguments{
\item{obj}{if you want to pass a existing `gp` instance to the app}
}
\description{
Just run gp_app() to start the Shiny-App in your RStudio or
browser
}
\examples{
library(dplyr)
p <- new.gp() \%>\%
  add_data(data.frame(x1 = 1:10, x2 = 2:11), y = sin(3:12))
gp_app(p)

}
