% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{help_opt}
\alias{help_opt}
\title{help_opt
Helperfunction which return the negative log likelihood for the selected parameter.}
\usage{
help_opt(par_value, X, par_names)
}
\arguments{
\item{par_value}{Value of the parameters, named list.}

\item{X}{X R6 object of the class "gp".}

\item{par_names}{Names of the parameters, character vecotr.}
}
\value{
Negative log likelihood function.
}
\description{
help_opt
Helperfunction which return the negative log likelihood for the selected parameter.
}
\examples{
T <- gp$new()
T$add_data(x = 1:10, y = sin(1:10))
help_opt(par_values = list(sigma = 0,
                           l = 1,
                           gamma = 1,
                           alpha = 1,
                           sigma0 = 1),
                       T,
                       par_names = c("sigma"))
}
