% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{init_cov}
\alias{init_cov}
\title{Creating covariance function, that does not depend on parameters}
\usage{
init_cov(covname, sigma = 0, l = 1, alpha = 1, sigma0 = 1, gamma = 1)
}
\arguments{
\item{covname}{Character vector, that expresses the name of the used
covariance function. One of "constant", "linear", "squared_exp", "exponential",
 "rational" or "gamma"}

\item{sigma}{numerical vector, length has to fit to the input dimension of
the used data}

\item{l}{numerical vector, has to be a non-negative scalar}

\item{alpha}{numerical vector, has to be a scalar}

\item{sigma0}{numerical vector, has to be a scalar}

\item{gamma}{numerical vector, has to be a scalar between 0 and 2}
}
\value{
covariance function of type 'closure'
}
\description{
Creating covariance function, that does not depend on parameters
}
\examples{
cov_fun <- init_cov("linear", sigma = c(1,2,3))
cov_fun(c(1,1,1), c(2,3,4))

}
