% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{linear_cov}
\alias{linear_cov}
\title{linear_cov}
\usage{
linear_cov(x, y, sigma)
}
\arguments{
\item{x}{numerical vector}

\item{y}{numerical vector}

\item{sigma}{numerical vector of the same length as x and y}
}
\value{
linear covariance of input parameters
}
\description{
linear_cov
}
\examples{
linear_cov(c(1,1), c(2,1), c(3,4))
}
