% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{new.gp}
\alias{new.gp}
\title{initializing a new gp-instance}
\usage{
new.gp(cov_fun = "squared_exp")
}
\arguments{
\item{cov_fun}{character vector, name of the covariance function, that should
be used. You have the choice between:
"constant", "linear", "squared_exp", "exponential", "gamma_exp", "rational"}
}
\value{
a "gp" object
}
\description{
initializing a new gp-instance
}
\examples{
p1 <- new.gp()
add_data(p1, X_learn = 1:10, y = 2:11)
print(p1)

}
