% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{optimize_gp}
\alias{optimize_gp}
\title{Optimizing hyperparameters of gp instance}
\usage{
optimize_gp(obj)
}
\arguments{
\item{obj}{a instance of class "gp"}
}
\value{
invisible return of list of optimized parameters for each covariance
function
}
\description{
Optimizing hyperparameters of gp instance
}
\details{
Optimization based on the maximization of the marginal likelihood.
}
\examples{
p1 <- new.gp()
add_data(p1, 1:5, 1:5)
plot(p1, 0,7)
optimize_gp(p1)
plot(p1, 0,7)

}
