% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{optimize_parameters}
\alias{optimize_parameters}
\title{optimize_parameters}
\usage{
optimize_parameters(X)
}
\arguments{
\item{X}{R6 object of the class "gp".}
}
\value{
List of values of parameters of each covariance function that minimize
the negative log marginal likelihood.
}
\description{
optimize_parameters
}
\examples{
T <- gp$new()
T$add_data(x = 1:10, y = sin(1:10))
T$optim_parameter()
}
