% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gp}
\alias{plot_gp}
\title{plotting a gp instance}
\usage{
plot_gp(gp, x_min = -5, x_max = 5, n_points = 100, plotly_obj = FALSE)
}
\arguments{
\item{gp}{gp instance}

\item{x_min}{start value x_axis}

\item{x_max}{end value x-axiy}

\item{n_points}{number of points generated}

\item{plotly_obj}{plotly object or not}
}
\description{
plotting a gp instance
}
