% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gp_posterior}
\alias{plot_gp_posterior}
\title{Plotting a sample plot}
\usage{
plot_gp_posterior(
  gp,
  x_min = -5,
  x_max = 5,
  n_points = 100,
  n_samples = 5,
  plotly_obj = FALSE
)
}
\arguments{
\item{gp}{gp instance}

\item{x_min}{start value x_axis}

\item{x_max}{end value x-axiy}

\item{n_points}{number of points generated}

\item{n_samples}{number of samples}

\item{plotly_obj}{plotly object or not}
}
\description{
Plotting a sample plot
}
