% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_laplace.R
\name{pred_laplace}
\alias{pred_laplace}
\title{calculating the prediction probability}
\usage{
pred_laplace(f_mode, X_learn, y_learn, cov_fun, likelihood_fun, x_input)
}
\arguments{
\item{f_mode}{values of the mode function, it must have the same length as y_learn}

\item{X_learn}{input values of the learning data, it's allowed to be a list
of points, a matrix or a data.frame. The number of represented points has to
equal the length of y_learn}

\item{y_learn}{output values of the learning data, it has to be a numeric vector
with 1 and -1/0 as values}

\item{cov_fun}{function, such that two numeric vectors x,y of the same length will
cause a scalar output}

\item{likelihood_fun}{name of the used function, a character vector, either 'probit'
or 'logit'}

\item{x_input}{a numeric vector of the same dimension as every point in X_learn,
point, where we want to predict the class label}
}
\value{
numeric vector, that describes the probability of the class_label 1 at point x_input
}
\description{
calculating the prediction probability
}
\examples{
pred_laplace(f_mode = c(1,2),
             X_learn = c(0,2),
             y_learn = c(0,1),
             cov_fun = function(x,y) exp(-abs(x*y)),
             likelihood_fun = "probit",
             x_input = 1)
-----------------------------------------------
pred_laplace(f_mode = c(1,2),
             X_learn = list(c(0,2), c(1,3)),
             y_learn = c(0,1),
             cov_fun = function(x,y) exp(-abs(x\%*\%y)),
             likelihood_fun = "probit",
             x_input = c(1,0))

}
