% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_laplace.R
\name{pred_mc_laplace}
\alias{pred_mc_laplace}
\title{Predictions using multiclass-laplace-approximations}
\usage{
pred_mc_laplace(X_learn, K_list, f_mode, cov_list, x_input, n_sample = 1000)
}
\arguments{
\item{X_learn}{list of input points}

\item{K_list}{list of variance-covariance-matrices}

\item{f_mode}{numeric vector of mode function values}

\item{cov_list}{list of closures, that fullfill the characteristics of a
covariance function}

\item{x_input}{numeric vector, coordinates of the input, where we want to get
the prediction}

\item{n_sample}{number of samples to approximate the probability of the class
labels}
}
\value{
vector describing the probabilities of the different labels
}
\description{
Predictions using multiclass-laplace-approximations
}
\examples{
x <- 1:50
y <-as.integer(x<5)
y <- c ( y, as.integer(x>5))
X_learn <- as.list(x)
covariance_list <- list(
  function(x,y) exp(- 0.5*sqrt(sum((x-y)^2))),
  function(x,y) exp(- 0.5*sqrt(sum((x-y)^2)))
)

K <- list(
 cov_cross(X_learn, X_learn,covariance_list[[1]]),
 cov_cross(X_learn, X_learn,covariance_list[[2]])
)

f_mode <- find_mode_mc_laplace(K,y)$mode

pred_mc_laplace(X_learn, K, f_mode, covariance_list, 20)
}
