% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_2.1.R
\name{predict_gauss}
\alias{predict_gauss}
\title{predict_gauss}
\usage{
predict_gauss(X_learn, y_learn, cov, noise, x_input, mean_fun = 0, ...)
}
\arguments{
\item{X_learn}{The learning data, given as an matrix, a list of vectors or a tibble}

\item{y_learn}{The values of the unknown function at the positions given in X_learn}

\item{cov}{The covariance-function, which is used. It has to provide a function call
like cov(x,y) with x and y vectors of the same size and outputs a scalar.}

\item{noise}{The assumed value of the variance of the noise of the learning data}

\item{x_input}{The position we want to predict the value of the unknown function}

\item{...}{further information to pass to covariance function. They have to be named}
}
\value{
list of three elements f_predict var_f log_marginal_likelihood
 f_predict The predicted value
 var_f The variance
}
\description{
predict_gauss
}
\examples{
predict_gauss(  X_learn = 1:10,
                y_learn = sin(1:10) + 1:10 + rnorm(10),
                cov = function(x,y) exp(-x*y),
                noise = 1,
                x_input = 12,
                mean_fun = function(x) x
)


}
