% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_laplace.R
\name{predict_laplace}
\alias{predict_laplace}
\title{using laplace-approximation for classification problems using gp}
\usage{
predict_laplace(X_learn, y_learn, cov_fun, likelihood_fun, x_input)
}
\arguments{
\item{X_learn}{input values of the learning data, it's allowed to be a list
of points, a matrix or a data.frame. The number of represented points has to
equal the length of y_learn}

\item{y_learn}{output values of the learning data, it has to be a numeric vector
with 1 and -1 as values}

\item{cov_fun}{function, such that two numeric vectors x,y of the same length will
cause a scalar output}

\item{likelihood_fun}{name of the used function, a character vector, either 'probit'
or 'logit'}

\item{x_input}{a numeric vector of the same dimension as every point in X_learn,
point, where we want to predict the class label}
}
\value{
numeric vector, that describes the probability of the class_label 1 at point x_input
}
\description{
using laplace-approximation for classification problems using gp
}
\examples{
predict_laplace(X_learn = c(0,2),
                y_learn = c(-1,1),
                cov_fun = function(x,y) exp(-sqrt(sum((x-y)^2))),
                likelihood_fun = "probit",
                x_input = 1)
------------------------------------------------------------
predict_laplace( X_learn = list(c(0,2), c(1,3)),
                y_learn = c(-1,1),
                cov_fun = function(x,y) exp(-sqrt(sum((x-y)^2))),
                likelihood_fun = "probit",
                x_input = c(1,0))
}
