% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{rational_quadratic_cov}
\alias{rational_quadratic_cov}
\title{rational_quadratic_cov}
\usage{
rational_quadratic_cov(x, y, l, alpha)
}
\arguments{
\item{x}{numerical vector}

\item{y}{numerical vector}

\item{l}{positive double value}

\item{alpha}{positive double value}
}
\value{
rational quadratic covariance of input parameters
}
\description{
rational_quadratic_cov
}
\examples{
rational_quadratic_cov(c(1,1),c(2,1), 1.5, 3)
}
