% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{set_cov}
\alias{set_cov}
\title{Setting covariance function}
\usage{
set_cov(obj, cov_fun)
}
\arguments{
\item{obj}{a instance of class "gp"}

\item{cov_fun}{a character. It has to be one of the following strings:
"constant", "linear", "squared_exp", "exponential","gamma_exp", "rational"}
}
\value{
invisible return of the modified object. Be aware, that all
setter-functions work on reference.
}
\description{
Setting covariance function
}
\details{
If you are interested in additional information about the different
covariance functions, you can visit the vignette of this package.
The main idea of those covariance functions is to describe, how the data
we already know is influencing the prediction of the value at some other
place.
If you are interested in optimizing the choice of the hyperparameter, try out
optimize_gp().
}
\examples{

p1 <- new.gp()
# the default value
get_cov(p1)

#modifing the value
set_cov(p1,"gamma_exp")
get_cov(p1)


}
