% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{set_mean_fun}
\alias{set_mean_fun}
\title{Adding a mean function to a gp object}
\usage{
set_mean_fun(obj, mean_fun)
}
\arguments{
\item{obj}{a instance of class "gp"}

\item{mean_fun}{either a numeric value or a closure, that just takes one input vector
and generates one-dimensional numeric output}
}
\value{
invisible return of the modified object. Be aware, that all
setter-functions work on reference.
}
\description{
Adding a mean function to a gp object
}
\examples{
p <- new.gp()
add_data(p, 1:10,1:10)
set_mean_fun(p, function(x) x)

}
