% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{set_noise}
\alias{set_noise}
\title{Setting the standard deviation of a gp instance}
\usage{
set_noise(obj, noise)
}
\arguments{
\item{obj}{a instance of class "gp"}

\item{noise}{a non-negative numerical value, that describes the standard deviation
of the noise}
}
\value{
invisible return of the modified object. Be aware, that all
setter-functions work on reference.
}
\description{
Setting the standard deviation of a gp instance
}
\examples{
p1 <- new.gp()
# the default value
get_noise(p1)

#modifing the vallue
set_noise(p1,2)
get_noise(p1)


}
