% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{set_parameter}
\alias{set_parameter}
\title{Setting the parameters of the used covariance function}
\usage{
set_parameter(
  obj,
  sigma = NULL,
  l = NULL,
  alpha = NULL,
  sigma0 = NULL,
  gamma = NULL
)
}
\arguments{
\item{obj}{a instance of class "gp"}

\item{sigma}{either a numerical vector of length input_dim or
a named list of numerical vectors, such that the names of the list elements are
a subset of the names of the parameters settable by this function}

\item{l}{a numerical value greater than zero}

\item{alpha}{a numerical value greater than zero}

\item{sigma0}{a numerical value}

\item{gamma}{a numerical value between 0 and 2}
}
\value{
invisible return of the modified object. Be aware, that all
setter-functions work on reference.
}
\description{
Setting the parameters of the used covariance function
}
\details{
Those parameters can be used to modify the used covariance functions.
Here is a list, which covariance function refers to which parameter
constant: sigma0
linear: sigma
squared_exp: l
exponential: l
rational: l, alpha
gamma_exp: l, gamma
For further information about the influence of each parameter, check out
the vignette of this package.
If you are interested in optimizing the choice of the hyperparameter, try out
optimize_gp().
}
\examples{
p <- new.gp()
add_data(p, 1:10,1:10)
get_parameter(p)    #default-values

set_parameter(p, sigma = 12, gamma = 1.5)
get_paramter(p)   #modified values

#updating via list
par_list <- list( sigma0 = 3, alpha = 2, l = 13)
set_parameter(p, par_list)
get_parameter(p)


}
