% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_functions.R
\name{squared_exp_cov}
\alias{squared_exp_cov}
\title{squared_exp_cov}
\usage{
squared_exp_cov(x, y, l)
}
\arguments{
\item{x}{numerical vector}

\item{y}{numerical vector}

\item{l}{positive double value}
}
\value{
squared exponential covariance of input parameters
}
\description{
squared_exp_cov
}
\examples{
squared_exp_cov(c(1,1),c(2,1), 4)
}
