% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_2.1.R
\name{cov_cross}
\alias{cov_cross}
\title{cov_cross}
\usage{
cov_cross(x, y, cov)
}
\arguments{
\item{x}{first list of input vectors}

\item{y}{second list of input vectors}

\item{cov}{covariance function, that returns a scalar}
}
\value{
variance-covariance-matrix
}
\description{
cov_cross
}
