% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_laplace.R
\name{find_mode_laplace}
\alias{find_mode_laplace}
\title{finding value of mode function using laplace-approximation}
\usage{
find_mode_laplace(K, y, likelihood_fun)
}
\arguments{
\item{K}{covariance-matrix of dimension (n,n)}

\item{y}{output-vector of length n, it's values are just allowed to be -1 or 1}

\item{likelihood_fun}{character-vector, name of the response function,
either "probit" or "logit"}
}
\value{
f_mode: value of mode function,
         log_marginal_likelihood: logarithmic marginal likelihood of the result
}
\description{
finding value of mode function using laplace-approximation
}
