% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{help_opt}
\alias{help_opt}
\title{help_opt
Helperfunction which return the negative log likelihood for the selected parameter.}
\usage{
help_opt(par_value, X, par_names)
}
\arguments{
\item{par_value}{Value of the parameters, named list.}

\item{X}{X R6 object of the class "gp".}

\item{par_names}{Names of the parameters, character vecotr.}
}
\value{
Negative log likelihood function.
}
\description{
help_opt
Helperfunction which return the negative log likelihood for the selected parameter.
}
