% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R, R/structure.R
\name{init_cov}
\alias{init_cov}
\title{#' Define different covariance functions.
#'
#' @param x Input data, a numeric vector.
#' @param y Input data, a numeric vector.
#' @param par The parameters of the covariance function used, a named list.
#'
#' @return value of the selected covariance function.
#' @export
linear <- function(x,y,par){
  linear_cov(x,y,par$sigma)
}
squared_exp <- function(x,y, par){
  squared_exp_cov(x,y,par$l)
}
constant <- function(x,y, par){
  constant_cov(x,y,par$sigma0)
}
exponential <- function(x,y, par){
  exp_cov(x,y,par$l)
}
gamma_exp <- function(x,y,par){
  gamma_exp_cov(x,y,par$l,par$gamma)
}
rational <- function(x,y,par){
  rational_quadratic_cov(x,y,par$l,par$alpha)
}}
\usage{
init_cov(covname, sigma = 0, l = 1, alpha = 1, sigma0 = 1, gamma = 1)

init_cov(covname, sigma = 0, l = 1, alpha = 1, sigma0 = 1, gamma = 1)
}
\arguments{
\item{covname}{Character vector, that expresses the name of the used
covariance function. One of "constant", "linear", "squared_exp", "exponential",
 "rational" or "gamma"}

\item{sigma}{numerical vector, length has to fit to the input dimension of
the used data}

\item{l}{numerical vector, has to be a non-negative scalar}

\item{alpha}{numerical vector, has to be a scalar}

\item{sigma0}{numerical vector, has to be a scalar}

\item{gamma}{numerical vector, has to be a scalar between 0 and 2}
}
\value{
The, with "covname", selected covariance function or if specified,
the value of the covariance function

covariance function of type 'closure'
}
\description{
cov_list <- list(
  "linear"= list(
    name = "linear",
    fun = linear,
    parameter = c("sigma"),
    upper = NULL,
    lower = NULL
  ),
}
\details{
"squared_exp"= list(
    name = "squared_exp",
    fun = squared_exp,
    parameter = c("l"),
    upper = NULL,
    lower = 0.0001
  ),

  "constant" = list(
    name = "constant",
    fun = constant,
    parameter = c("sigma0"),
    upper = NULL,
    lower = NULL
  ),
  "exponential" = list(
    name = "exponential",
    fun = exponential,
    parameter = c("l"),
    upper = NULL,
    lower = 0.0001
  ),
  "gamma_exp" = list(
    name = "gamma_exp",
    fun = gamma_exp,
    parameter = c("l","gamma"),
    upper = c(Inf, 2),
    lower = c(0.0001, 0.01)
  ),
  "rational" =list(
    name = "rational",
    fun = rational,
    parameter = c("l", "alpha"),
    upper = c(Inf, Inf),
    lower = c(0.0001, 0.0001)
  )
)

init_cov
}
\examples{
init_cov("linear", sigma = 2)(x = 1:10, y = sin(1:10))
t <- init_cov("rational", l = 1, alpha = 0.5)
t(x = 1:5, y = exp(1:5))
cov_fun <- init_cov("linear", sigma = c(1,2,3))
cov_fun(c(1,1,1), c(2,3,4))

}
