% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{optimize_gp}
\alias{optimize_gp}
\title{Optimizing hyperparameters of gp instance}
\usage{
optimize_gp(obj)
}
\arguments{
\item{obj}{a instance of class "gp"}
}
\value{
invisible return of the modified object.
}
\description{
Optimizing hyperparameters of gp instance
}
\details{
Further details about the used method.
Optimiziation based on the maximization of the marginal likelihood.
}
